fix: set authentification token in headers after login
Previously, when the user clicks the 'confirm' button of the login form :

1. The login async thunk function was called. A post call using axios set
req.headers.authorization to localStorage.getItem('accessToken'), which
was not yet set.
2. The login reducer set localStorage.setItem('accessToken') to the
axios call's response.accessToken.

This resulted in the access Token being undefined when sent to the back end.

Now, the accessToken is set directly from the axios call's response data
and set properly to the backend, which makes the authentification middle-
ware work.